Pass state as props to Spinner
Spinner is either shown entirely or can be hidden internally. It must be able to handle show/hide logic internally because its parent, in this case IdeaCreate, may need to avoid re-rendering because its listening for an event emitted by an element (an iFrame). That event would endlessly be restarted otherwise.